# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GSI Partner Opportunity Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /internal-resources:
    get:
      operationId: listInternalResources
      tags: [internalResources]
      summary: List internal people resources
      responses:
        "200":
          description: List of internal resources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InternalResource"
    post:
      operationId: createInternalResource
      tags: [internalResources]
      summary: Create an internal resource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InternalResourceInput"
      responses:
        "201":
          description: Created internal resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalResource"
        "200":
          description: Existing internal resource reused (same normalized name+email)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalResource"
  /internal-resources/bulk:
    post:
      operationId: bulkImportInternalResources
      tags: [internalResources]
      summary: Bulk-import contacts into the internal people directory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkInternalContactImportBody"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkContactImportResult"
        "400":
          description: Invalid request
  /internal-resources/{id}:
    patch:
      operationId: updateInternalResource
      tags: [internalResources]
      summary: Update internal resource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InternalResourceUpdate"
      responses:
        "200":
          description: Updated internal resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalResource"
        "404":
          description: Not found
    delete:
      operationId: deleteInternalResource
      tags: [internalResources]
      summary: Delete internal resource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # REMINDERS
...
  /opportunities/{id}/meddpicc-sections:
    get:
      operationId: getMeddpiccSections
      tags: [meddpicc]
      summary: Get per-element notes and associated contact ids for an opportunity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Section meta for each element that has any
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MeddpiccSectionMeta"
  /opportunities/{id}/meddpicc-sections/{element}:
    put:
      operationId: updateMeddpiccSection
      tags: [meddpicc]
      summary: Upsert notes and associated contact ids for a single MEDDPICC element
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: element
          in: path
          required: true
          schema:
            type: string
            enum: [metrics, economic_buyer, decision_criteria, decision_process, paper_process, identify_pain, champion, competition]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeddpiccSectionMetaUpdate"
      responses:
        "200":
          description: Updated section meta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeddpiccSectionMeta"
...
  /backup/export:
    get:
      operationId: exportBackup
      tags: [backup]
      summary: Export a full JSON backup of all user-entered data (admin only)
      responses:
        "200":
          description: Full database backup
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackupData"
        "403":
          description: Admin access required
  /backup/import:
    post:
      operationId: importBackup
      tags: [backup]
      summary: Restore the database from a JSON backup, replacing all data (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BackupData"
      responses:
        "200":
          description: Restore result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackupActionResult"
        "400":
          description: Invalid or unsafe backup payload
        "403":
          description: Admin access required
  /backup/clear:
    post:
      operationId: clearData
      tags: [backup]
      summary: Delete all user-entered data, preserving only the admin credentials (admin only)
      responses:
        "200":
          description: Clear result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackupActionResult"
        "403":
          description: Admin access required
components:
  schemas:
...
    ContactImportRow:
      type: object
      required: [name]
      properties:
        name:
          type: string
        func:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        isManager:
          type: boolean
        notes:
          type: string
          nullable: true
          description: Free-text note; converted to a single Note entry on the server
...
    BulkInternalContactImportBody:
      type: object
      required: [rows]
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/ContactImportRow"
    BulkContactImportResult:
      type: object
      required: [inserted, skipped, errors]
      properties:
        inserted:
          type: integer
        skipped:
          type: integer
        errors:
          type: array
          items:
            type: string
    BackupData:
      type: object
      required: [tables]
      properties:
        version:
          type: string
        exportedAt:
          type: string
        tables:
          type: object
          additionalProperties:
            type: array
            items:
              type: object
              additionalProperties: true
    BackupActionResult:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
        message:
          type: string
        counts:
          type: object
          additionalProperties:
            type: integer
...
    InternalResource:
      type: object
      required: [id, name, isManager, notes, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        func:
          type: ["string", "null"]
          description: The person's function / role in the company
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        isManager:
          type: boolean
        managerId:
          type: ["integer", "null"]
        notes:
          type: array
          items:
            $ref: "#/components/schemas/Note"
        createdAt:
          type: string
    InternalResourceInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        func:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        isManager:
          type: boolean
        managerId:
          type: ["integer", "null"]
        notes:
          type: array
          items:
            $ref: "#/components/schemas/Note"
    InternalResourceUpdate:
      type: object
      properties:
        name:
          type: string
        func:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        isManager:
          type: boolean
        managerId:
          type: ["integer", "null"]
        notes:
          type: array
          items:
            $ref: "#/components/schemas/Note"

    # PARTNER RESOURCE
...
    Note:
      type: object
      required: [id, text, createdAt]
      properties:
        id:
          type: string
        text:
          type: string
        createdAt:
          type: string
...
    MeddpiccElement:
      type: string
      enum: [metrics, economic_buyer, decision_criteria, decision_process, paper_process, identify_pain, champion, competition]
...
    MeddpiccSectionMeta:
      type: object
      required: [element, notes, contactIds]
      properties:
        element:
          $ref: "#/components/schemas/MeddpiccElement"
        notes:
          type: array
          items:
            $ref: "#/components/schemas/Note"
        contactIds:
          type: array
          items:
            type: string
        ownerId:
          type: string
          nullable: true
    MeddpiccSectionMetaUpdate:
      type: object
      properties:
        notes:
          type: array
          items:
            $ref: "#/components/schemas/Note"
        contactIds:
          type: array
          items:
            type: string
        ownerId:
          type: string
          nullable: true
...