# lib/api-zod/src/generated/types/internalResource.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GSI Partner Opportunity Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { Note } from './note';

export interface InternalResource {
  id: number;
  name: string;
  /**
     * The person's function / role in the company
     * @nullable
     */
  func?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  location?: string | null;
  isManager: boolean;
  /** @nullable */
  managerId?: number | null;
  notes: Note[];
  createdAt: string;
}