# lib/api-zod/src/generated/types/internalResourceUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GSI Partner Opportunity Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { Note } from './note';

export interface InternalResourceUpdate {
  name?: string;
  /** @nullable */
  func?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  location?: string | null;
  isManager?: boolean;
  /** @nullable */
  managerId?: number | null;
  notes?: Note[];
}